remote_llm_server/requirements.txt
llama-cpp-python[server]==0.2.90
uvicorn
fastapi
sse-starlette
pydantic-settings
